performance: speed up ignore comment detection
Tokenizing a file is quite slow, as the module is implemented in pure python. So, instead of always tokenizing the files when use_noqa is set, we will first use a fast regex.

The following command on the LibCST repo went from 26 seconds to 17 seconds.
fixit run_rules --rules RewriteToLiteralRule UseAssertInRule UseFstringRule NoRedundantFStringRule