Enable FullRepoManager when lint rules need it
Looks at the list of enable lint rules, and if any of those need the
`FullyQualifiedNameProvider` metadata, and then enables the
`FullRepoManager` from libcst. The cache is then resolved and passed
through the metadata wrapper when visiting the CST, enabling rules to
use the appropriate metadata.

Requires passing the root path and file path being linted to
`FullRepoManager`, which is accomplished by passing the Fixit config
further through the API into the `CSTLintRunner` implementation.

[ghstack-poisoned]