Make `VALID` and `INVALID` covariant
By using `Sequence`, these become covariant. Previously, doing something like:
```
    VALID = [
        ValidTestCase("..."),
        ValidTestCase("..."),
    ]
```

would be a type error in `pyre`, because `pyre` would infer this to be a `list[ValidTestCase]` which is *not* a subclass of `list[str | ValidTestCase]`. By making these `Sequence`s (and thus immutable), we can avoid the requirement of having explicit type annotations everywhere for this.